Revert "Lpb page unfinished"
